+++ client/src/lib/Advisories/Advisory.svelte
@@ -16,7 +16,7 @@
   import { convertToDocModel } from "$lib/Advisories/CSAFWebview/docmodel/docmodel";
   import SsvcCalculator from "$lib/Advisories/SSVC/SSVCCalculator.svelte";
   import { convertVectorToLabel } from "$lib/Advisories/SSVC/SSVCCalculator";
-  import JsonDiff from "$lib/Diff/JsonDiff.svelte";
+  import Diff from "$lib/Diff/Diff.svelte";
   import { ARCHIVED, ASSESSING, DELETE, NEW, READ, REVIEW } from "$lib/workflow";
   import { canSetStateRead } from "$lib/permissions";
   import CommentTextArea from "./Comments/CommentTextArea.svelte";
@@ -61,7 +61,6 @@
   }
 
   const setAsReadTimeout: number[] = [];
-  let diffDocuments: any;
   let isDiffOpen = false;
   let commentFocus = false;
 
@@ -312,14 +311,6 @@
     }
   }
 
-  const onSelectedDiffDocuments = async (event: any) => {
-    diffDocuments = {
-      docA: event.detail.docA,
-      docB: event.detail.docB
-    };
-    isDiffOpen = true;
-  };
-
   onDestroy(() => {
     setAsReadTimeout.forEach((id: number) => {
       clearTimeout(id);
@@ -366,14 +357,14 @@
               trackingID={params.trackingID}
               {advisoryVersions}
               selectedDocumentVersion={document.tracking?.version}
-              on:selectedDiffDocuments={onSelectedDiffDocuments}
+              on:selectedDiffDocuments={() => (isDiffOpen = true)}
               on:disableDiff={() => (isDiffOpen = false)}
             ></Version>
           {/if}
         </div>
         <div class="flex flex-col">
           {#if isDiffOpen}
-            <JsonDiff title={undefined} {diffDocuments}></JsonDiff>
+            <Diff showTitle={false}></Diff>
           {:else}
             <Webview></Webview>
           {/if}
